requirements.txt
ucloud-sdk-python3
pyyaml
